Remove singleton
The singleton (`cf`) in clientfactory was problematic from a
testing point of view (we could not run tests in parallel).

This patch removes it by adding a private `model.Submodel`
to the SubProvider.

This will allow us to update code to pass a per-SubProvider
instance client factory to resources via the SubProvider's
GetResources/GetDataSources calls. (Code in a follow on patch.)

We also update the SubProvider interface to add `contexts`
(so they can be used for logging if desired).