Add some changes to client behaviour (#27)
We now no longer have a wrapper around the morpheus
client, instead we work directly with a standard
morpheus API client (`client.APIClient`).

Rather than having a wrapper that implements an auth
roundtripper, we instead insert an auth roundtripper into the
http client used by the morpheus client. This should allow
us to work with http transports in the usual way.

Seperate username/password and token auth roundtrippers.
(This simplifies the token case).

Provide the ability to pass a custom http client to the
client factory - this is mainly intended for test purposes.

No longer generate a token when we instantiate a client; we
will now generate a token only when it is needed.