Add role_type field to generalised role (#65)
Adds the ability for the user to set the "role_type" field.

This determines whether a User Role or Account Role is created in
Morpheus.

We're adding the role_type field now to prevent any potential state file
oddities.

As per feedback from Eamonn, I'm currently only supporting user roles
with the first provider release. The provider will exit with an error
message if the user tries to provide an "account" role_type, and no role
will be created.