user resource: write only password (#69)
Switch to a write only password. This `password_wo` is not
persisted in the state file.

We also introduce `password_wo_version` which can be used (later) in
Update()
to determine if the password has been changed. (This is required because
terraform cannot determine if the write-only password has changed by
examining the state.)