user resource: add plan modifier (#74)
This plan modifier ensures that if `password_wo` is missing during
plan it is detected and an error is raised.

`password_wo` is a corner case where it is required for create, but
not for import. We do not raise an error in the import case
(tested by existing test).