user resource: add import test (#75)
We are already testing import of user resource, but here we add
a new test which adds an alternative pattern.

We perform two phases:

phase 1:

* uses terraform to create a user
* removes that user from terraform control
* does not delete that user

phase 2:

* inherits no state from phase 1
* performs import in a way that the state is preserved post import
* tests performed post import will inherit (and potentially modify) the
import state

Some parts of this were tricky, because the test
framework makes some assumptions that are not
the case when using write_only attributes.

We are able to test create with password_wo
specified, and import without it specified --
while preserving the import state for follow
on tests.